requirements.txt
langchain
python-dotenv
langchain_nvidia_ai_endpoints
langchain_community
streamlit
faiss-cpu
pypdf